refactor: make _api.py a coherent user-facing module
Move misplaced functions out of _api.py to their proper modules:

- cleanup_and_store, acleanup_and_store, scrub_run_output_for_storage
  → _storage.py (post-run persistence orchestration)
- update_reasoning_content_from_tool_call
  → _response.py (reasoning/response processing)
- get_effective_filters
  → _messages.py (knowledge filter resolution)

_api.py now only contains user-facing APIs: print_response and cli_app.
Updated all delegation calls in agent.py and _hooks.py accordingly.
